requirements.txt
discord.py
pywin32
youtube_dl
python-dotenv